pyproject.toml
[project]
name = "cost-estimation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.75",
    "langgraph>=0.6.6",
    "openai>=1.105.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
